+++ src/MoBi.UI/Views/InitialConditionsView.cs
@@ -18,4 +18,4 @@ namespace MoBi.UI.Views
          _presenter = presenter;
       }
    }
-}
+}i
